+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Unreleased
 
 - fix: reject create conflicts and duplicate permission targets
+- fix(sql): seed ns export types coredns & native for 2.x upgrade
 
 ### [3.0.3] - 2026-07-27